+++ python/ray/train/tests/test_gpu_examples.py
@@ -1,8 +1,6 @@
 import pytest
 import torch
 
-import ray
-from ray.cluster_utils import Cluster
 from ray.air import Checkpoint, session
 
 from ray.air.config import ScalingConfig
@@ -23,41 +21,6 @@ from ray.train.tests.test_tune import (
 )
 from ray.train.tensorflow.tensorflow_trainer import TensorflowTrainer
 from ray.train.torch.torch_trainer import TorchTrainer
-
-
-@pytest.fixture
-def ray_start_4_cpus_2_gpus():
-    address_info = ray.init(num_cpus=4, num_gpus=2)
-    yield address_info
-    # The code after the yield will run as teardown code.
-    ray.shutdown()
-
-
-@pytest.fixture
-def ray_start_1_cpu_1_gpu():
-    address_info = ray.init(num_cpus=1, num_gpus=1)
-    yield address_info
-    ray.shutdown()
-
-
-@pytest.fixture
-def shutdown_only():
-    yield None
-    ray.shutdown()
-
-
-@pytest.fixture
-def ray_2_node_2_gpu():
-    cluster = Cluster()
-    for _ in range(2):
-        cluster.add_node(num_cpus=4, num_gpus=2)
-
-    ray.init(address=cluster.address)
-
-    yield
-
-    ray.shutdown()
-    cluster.shutdown()
 
 
 def test_tensorflow_mnist_gpu(ray_start_4_cpus_2_gpus):
